+++ packages/mint-components/src/components/sqm-referral-table/columns/sqm-referral-table-status-column.tsx
@@ -7,7 +7,7 @@ import { ReferralTableColumn } from "./ReferralTableColumn";
  * @uiName Referral Table Status Column
  * @validParents ["sqm-referral-table"]
  * @exampleGroup Referrals
- * @example Referral Table Status Column - <sqm-referral-table-status-column column-title="Status" converted-status-text="Converted" in-progress-status-text="In Progress"pending-review-status-text="Pending" denied-status-text="Denied"></sqm-referral-table-status-column>
+ * @example Referral Table Status Column - <sqm-referral-table-status-column column-title="Status" converted-status-text="Converted" in-progress-status-text="In Progress" pending-review-status-text="Pending" denied-status-text="Denied" pending-review-status-sub-text="Awaiting review" denied-status-sub-text="Detected self-referral"></sqm-referral-table-status-column>
  */
 @Component({
   tag: "sqm-referral-table-status-column",
@@ -39,6 +39,16 @@ export class ReferralTableStatusColumn implements ReferralTableColumn {
    */
   @Prop() deniedStatusText: string = "Denied";
 
+  /**
+   * @uiName Pending review status sub-text
+   */
+  @Prop() pendingReviewStatusSubText: string = "Awaiting review";
+
+  /**
+   * @uiName Denied status sub-text
+   */
+  @Prop() deniedStatusSubText: string = "Detected self-referral";
+
   constructor() {
     withHooks(this);
   }
@@ -48,14 +58,17 @@ export class ReferralTableStatusColumn implements ReferralTableColumn {
   async renderCell(data: Referral) {
     // TODO: Make ICU and more complete
     let statusText: string;
+    let statusSubText: string;
 
     const fraudStatus: "PENDING" | "DENIED" | "APPROVED" =
       data?.fraudData?.moderationStatus;
 
     if (fraudStatus === "DENIED") {
       statusText = this.deniedStatusText;
+      statusSubText = this.deniedStatusSubText;
     } else if (fraudStatus === "PENDING") {
       statusText = this.pendingReviewStatusText;
+      statusSubText = this.pendingReviewStatusSubText;
     } else {
       statusText = data.dateConverted
         ? this.convertedStatusText
@@ -65,6 +78,7 @@ export class ReferralTableStatusColumn implements ReferralTableColumn {
     return (
       <sqm-referral-table-status-cell
         status-text={statusText}
+        status-sub-text={statusSubText}
         fraud-status={fraudStatus}
         converted={data.dateConverted ? true : false}
       ></sqm-referral-table-status-cell>